pyproject.toml
[project]
name = "minbin"
version = "0.1.0"
description = "A lightweight self-hostable temporary Pastebin service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
]
